fix: Cookies will work when url is a $loopElement
The issue was that we were computing the url after setting the cookies
if it was templated. This led to the cookies being set on '' domain in
the cookie jar leading to it not being set on the request properly. The
fix is to move the logic of url computation before setting the cookies.
Tests are also added for the same so that we can detect if it would
break in future due to code changes.